django compatibility
more moving code around
