+++ mathics/doc/gather.py
@@ -7,9 +7,161 @@ Functions used to build the reference sections from module information.
 """
 
 
+import os.path as osp
 import pkgutil
+from os import listdir
 from types import ModuleType
 from typing import Tuple
+
+from mathics.core.util import IS_PYPY
+
+
+def filter_toplevel_modules(module_list):
+    """
+    Keep just the modules at the top level.
+    """
+    if len(module_list) == 0:
+        return module_list
+
+    modules_and_levels = sorted(
+        ((module.__name__.count("."), module) for module in module_list),
+        key=lambda x: x[0],
+    )
+    top_level = modules_and_levels[0][0]
+    return (entry[1] for entry in modules_and_levels if entry[0] == top_level)
+
+
+def gather_docs_from_files(documentation, path):
+    """
+    Load documentation from files in path
+    """
+    # First gather data from static XML-like files. This constitutes "Part 1" of the
+    # documentation.
+    files = listdir(path)
+    files.sort()
+
+    chapter_order = 0
+    for file in files:
+        part_title = file[2:]
+        if part_title.endswith(".mdoc"):
+            part_title = part_title[: -len(".mdoc")]
+            # If the filename start with a number, then is a main part. Otherwise
+            # is an appendix.
+            is_appendix = not file[0].isdigit()
+            chapter_order = documentation.load_part_from_file(
+                osp.join(path, file),
+                part_title,
+                chapter_order,
+                is_appendix,
+            )
+
+
+def gather_reference_part(documentation, title, modules, builtins_by_module):
+    """
+    Build a part from a title, a list of modules and information
+    of builtins by modules.
+    """
+    part_class = documentation.part_class
+    reference_part = part_class(documentation, title, True)
+    modules = filter_toplevel_modules(modules)
+    modules_seen = set([])
+    for module in sorted_modules(modules):
+        if skip_module_doc(module, modules_seen):
+            continue
+        chapter = documentation.doc_chapter(module, reference_part, builtins_by_module)
+        if chapter is None:
+            continue
+        reference_part.chapters.append(chapter)
+    return reference_part
+
+
+def doc_chapter(part, module, builtins_by_module):
+    """
+    Build documentation structure for a "Chapter" - reference section which
+    might be a Mathics Module.
+    """
+    # TODO: reformulate me in a way that symbols are always translated to
+    # sections, and guide sections do not contain subsections.
+
+    documentation = part.documentation
+    chapter_class = documentation.chapter_class
+    doc_class = documentation.doc_class
+
+    modules_seen = set([])
+    title, text = get_module_doc(module)
+    chapter = chapter_class(part, title, doc_class(text, title, None))
+    builtins = builtins_by_module.get(module.__name__)
+
+    if module.__file__.endswith("__init__.py"):
+        # We have a Guide Section.
+
+        # This is used to check if a symbol is not duplicated inside
+        # a guide.
+        submodule_names_seen = set([])
+        name = get_doc_name_from_module(module)
+        guide_section = documentation.add_section(
+            chapter, name, module, operator=None, is_guide=True
+        )
+        submodules = [
+            value for value in module.__dict__.values() if isinstance(value, ModuleType)
+        ]
+
+        # Add sections in the guide section...
+        for submodule in sorted_modules(submodules):
+            if skip_module_doc(submodule, modules_seen):
+                continue
+            elif IS_PYPY and submodule.__name__ == "builtins":
+                # PyPy seems to add this module on its own,
+                # but it is not something that can be importable
+                continue
+
+            submodule_name = get_doc_name_from_module(submodule)
+            if submodule_name in submodule_names_seen:
+                continue
+            section = documentation.add_section(
+                chapter,
+                submodule_name,
+                submodule,
+                operator=None,
+                is_guide=False,
+                in_guide=True,
+            )
+            modules_seen.add(submodule)
+            submodule_names_seen.add(submodule_name)
+            guide_section.subsections.append(section)
+
+            builtins = builtins_by_module.get(submodule.__name__, [])
+            subsections = list(builtins)
+            for instance in subsections:
+                if hasattr(instance, "no_doc") and instance.no_doc:
+                    continue
+
+                name = instance.get_name(short=True)
+                if name in submodule_names_seen:
+                    continue
+
+                submodule_names_seen.add(name)
+                modules_seen.add(instance)
+
+                documentation.add_subsection(
+                    chapter,
+                    section,
+                    name,
+                    instance,
+                    instance.get_operator(),
+                    in_guide=True,
+                )
+    else:
+        if not builtins:
+            return None
+        sections = [builtin for builtin in builtins if not skip_doc(builtin.__class__)]
+        documentation.doc_sections(sections, modules_seen, chapter)
+    return chapter
+
+
+def new_gather_sections(chapter, module, builtins_by_module) -> list:
+    """Build a list of DocSections from a "top-level" module"""
+    pass
 
 
 def get_module_doc(module: ModuleType) -> Tuple[str, str]:
@@ -100,4 +252,15 @@ def skip_module_doc(module, must_be_skipped) -> bool:
         or module.__name__.split(".")[0] not in ("mathics", "pymathics")
         or hasattr(module, "no_doc")
         and module.no_doc
+    )
+
+
+def sorted_modules(modules) -> list:
+    """Return modules sorted by the ``sort_order`` attribute if that
+    exists, or the module's name if not."""
+    return sorted(
+        modules,
+        key=lambda module: module.sort_order
+        if hasattr(module, "sort_order")
+        else module.__name__,
     )
